TEAL-546: Remove test_connection_to_invalid_database_password test case
While this test case passed without issue in a local development environment, it failed in Gitlab CI. The failure was caused because the expected exception (HTTPException) wasn't raised, telling us that the test inside Gitlab CI successfully connected to the database despite using the wrong password. A possible cause of the issue is that passwordless authentication to the postgres service is is somehow being enabled in Gitlab CI but this hasn't been verified. For now, the test case has been removed but this should be investigated in more detail in the future.