TEAL-558: Implement Bearer token authentication middleware skeleton
Reverts much of the previous commit (3f0031a TEAL-558: Authenticate OIDC tokens with authlib). The MALTOPUFT backend no longer interacts with SKA IAM directly. The login flow is carried out between the frontend and the auth-api (see TEAL-549). This results in the frontend session possesing an access token for the auth-api.

This access token is embedded in the request headers of all requests to the backend. This commit implements a middleware to exchange the auth-api access token for a maltopuft-backend access token and inject the user information into the request headers.

The token exchange is not actually implemented yet, so as a stand in the auth-api access token is just being decoded, which doesn't provide much in the way of security. This should therefore just be considered an authentication middlware skeleton.