DAAC-145: Refactor controllers to factory pattern
This refactor doesn't change the behaviour of any existing application code.

A key reason for the refactor was because the controllers were being instantiated as singletons (one object was instantiated and used throughout the application). This made it difficult to use a controller in a different sub-package without encountering issues with circular dependencies.

The Factory pattern gives a central location to import and create controllers which alleviates the dependency issues.