DAAC-96: Return one model instance in BaseRepository._execute_one
Previously, an object of type Row[Model] was returned. Accessing the model instance therefore required checking the length of the returned object and accessing the 0th index of the result to access the model instance. This change moves that logic into the _execute_one() method. This ensures that either one model instance or None is returned, which better matches the intended usage and naming of the method.