DAAC-96: Remove User.email attribute
It was always intended for the user information in maltopuftdb to be created from the SKA-IAM access token when a user first logged in to the platform. It turns out that the SKA-IAM access tokens don't actually include the user's email address so this attribute should not be present in the database table at all.

This commit also adds a User SQLAlchemy model repr.