DAAC-96: Embed database user information in requests
An additional method has been added to the AuthenticatedBackend which attempts to retrieve a user from maltopuftdb with the sub (user uuid) extracted from the bearer token. If the bearer token is a valid one and the user doesn't exist in the database, then it is assumed that the user is logging in for the first time, and the user record is created in maltopuftdb.